// apps/web/app/api/enquiry/models.ts
export type AddEnquiryModel = {
  name: string;
  date: string;
  emailId: string;
  contactNumber: string;
  dob: string;
  gender: string;
  age: string;
  maritalStatus: string;
  fatherName: string;
  fatherOccupation: string;
  motherName: string;
  motherOccupation: string;
  guardianName: string;
  guardianOccupation: string;
  addressLine1: string;
  addressLine2: string;
  nationality: string;
  state: string;
  district: string;
  postalCode: string;
  schoolName10th: string;
  obtainedMark10th: string;
  yearOfPassing10th: string;
  mediumOfEducation10th: string;
  schoolName12th: string;
  obtainedMark12th: string;
  yearOfPassing12th: string;
  mediumOfEducation12th: string;
};